requirements.txt
pandas>=2.0
polars>=0.20
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
plotly>=5.18
streamlit>=1.30
pytest>=7.4
pyyaml>=6.0
requests>=2.31
python-dotenv>=1.0
duckdb>=0.9
numpy>=1.24
